# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      last_login:
        type: string
      password:
        description: Hashed password
        type: string
      role:
        description: e.g., "admin", "user"
        type: string
      status:
        description: e.g., "active", "inactive"
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
host: localhost:3000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the API documentation for the Monitron Backend.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Monitron Backend API
  version: "1.0"
paths:
...
  /password/forgot:
    post:
      consumes:
      - application/json
      description: Sends a password reset link to the user\'s email
      parameters:
      - description: User email
        in: body
        name: email
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 'message": "If an account with that email exists, a password
            reset link has been sent.'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 'error": "Cannot parse JSON'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error": "Could not initiate password reset'
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Initiate password reset
      tags:
      - User Management
  /password/reset:
    post:
      consumes:
      - application/json
      description: Resets user password using a valid token
      parameters:
      - description: Token and new password
        in: body
        name: resetRequest
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 'message": "Password reset successfully'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 'error": "Invalid or expired token" or "Cannot parse JSON'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error": "Could not reset password'
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Reset password with token
      tags:
      - User Management
...
  /user/change-password:
    put:
      consumes:
      - application/json
      description: Allows an authenticated user to change their password
      parameters:
      - description: Current and new passwords
        in: body
        name: passwordChange
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 'message": "Password changed successfully'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 'error": "Cannot parse JSON'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 'error": "Invalid current password'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error": "Could not change password'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Change password for authenticated user
      tags:
      - User Management
  /users:
    get:
      description: Retrieve a list of all registered users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "500":
          description: 'error": "Could not retrieve users'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get all users (Admin Only)
      tags:
      - User Management
  /users/{id}:
    delete:
      description: Delete a user by their ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: 'error": "Invalid user ID'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 'error": "User not found'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error": "Could not delete user'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete a user (Admin Only)
      tags:
      - User Management
    get:
      description: Retrieve a single user by their ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: 'error": "Invalid user ID'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 'error": "User not found'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error": "Could not retrieve user'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get user by ID (Admin Only)
      tags:
      - User Management
    put:
      consumes:
      - application/json
      description: Update details of an existing user by their ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: User object with updated fields
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: 'error": "Invalid user ID" or "Cannot parse JSON'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 'error": "User not found'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error": "Could not update user'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update an existing user (Admin Only)
      tags:
      - User Management
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"